# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/internal/admin/ledger/rejections/{id}/resolve" : {
      "post" : {
        "tags" : [ "admin-ledger-rejection-controller" ],
        "operationId" : "resolve",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LedgerRejectionDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/ledger/reconciliation/runs/{id}/resolve-breaks" : {
      "post" : {
        "tags" : [ "reconciliation-controller" ],
        "operationId" : "resolveBreak",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResolveBreakRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/internal/admin/ledger/reconciliation/runs/{id}/execute" : {
      "post" : {
        "tags" : [ "reconciliation-controller" ],
        "operationId" : "executeRun",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "date",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationRun"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/ledger/reconciliation/runs" : {
      "get" : {
        "tags" : [ "reconciliation-controller" ],
        "operationId" : "getRuns",
        "parameters" : [ {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageReconciliationRun"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/ledger/reconciliation/runs/{id}" : {
      "get" : {
        "tags" : [ "reconciliation-controller" ],
        "operationId" : "getRun",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReconciliationRun"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/ledger/reconciliation/runs/{id}/breaks" : {
      "get" : {
        "tags" : [ "reconciliation-controller" ],
        "operationId" : "getBreaks",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "kind",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "GATEWAY_VS_LEDGER", "ORDERS_VS_CLEARING", "WALLET_VS_LEDGER", "PAYABLE_VS_ORDERS", "DOUBLE_ENTRY", "STUCK" ]
          }
        }, {
          "name" : "resolved",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "pageable",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageReconciliationBreak"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ResolveRequest" : {
        "type" : "object",
        "properties" : {
          "note" : {
            "type" : "string"
          }
        }
      },
      "LedgerRejectionDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "eventId" : {
            "type" : "string"
          },
          "producer" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "payload" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resolvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resolvedBy" : {
            "type" : "string"
          },
          "resolutionNote" : {
            "type" : "string"
          },
          "ageMinutes" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ResolveBreakRequest" : {
        "type" : "object",
        "properties" : {
          "resolvedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
      "ReconciliationRun" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "finishedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          }
        }
      },
...
      "PageableObject" : {
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "unpaged" ],
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        }
      },
      "SortObject" : {
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          }
        }
      },
      "PageReconciliationRun" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReconciliationRun"
            }
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "PageReconciliationBreak" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReconciliationBreak"
            }
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "ReconciliationBreak" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "runId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "GATEWAY_VS_LEDGER", "ORDERS_VS_CLEARING", "WALLET_VS_LEDGER", "PAYABLE_VS_ORDERS", "DOUBLE_ENTRY", "STUCK" ]
          },
          "subjectType" : {
            "type" : "string"
          },
          "subjectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "expected" : {
            "type" : "number"
          },
          "actual" : {
            "type" : "number"
          },
          "detail" : {
            "type" : "string"
          },
          "resolvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resolvedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "note" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}